Fixed #38: prevented divizion by zero
Some tracks recorded by devices which records with frequency > 1 Hz and
with time without fraction.
In example in this case we can have 5-10 points with the same time.
When calculating averages by distances time is used for division and division
by zero produced NaN's and Infinities, so later another methods that depends on
comparison Floats fails.